Update QA scripts. Main changes are:
The big crontab.txt is now split into 3 due to an Ubuntu bug:
https://bugs.launchpad.net/ubuntu/+source/cron/+bug/1972830

Calls to update_floods.sh have moved to root's crontab.
Not strictly necessary on Ubuntu 18.04 but needed on Debian 11
and Ubuntu 22.04.

The "generic" relation script handles multiple accounts; this
must run from root if these are different.